# feat(rule): add more perf timing metrics to rules