--- a/doc/API.md
+++ b/doc/API.md
@@ -338,19 +338,20 @@ The options parameter is flexible way to configure how `axe.run` operates. The d
 
 Additionally, there are a number or properties that allow configuration of different options:
 
-| Property        | Default | Description                                                                                                                             |
-| --------------- | :------ | :-------------------------------------------------------------------------------------------------------------------------------------- |
-| `runOnly`       | n/a     | Limit which rules are executed, based on names or tags                                                                                  |
-| `rules`         | n/a     | Allow customizing a rule's properties (including { enable: false })                                                                     |
-| `reporter`      | `v1`    | Which reporter to use (see [Configuration](#api-name-axeconfigure))                                                                     |
-| `resultTypes`   | n/a     | Limit which result types are processed and aggregated                                                                                   |
-| `xpath`         | `false` | Return xpath selectors for elements                                                                                                     |
-| `absolutePaths` | `false` | Use absolute paths when creating element selectors                                                                                      |
-| `iframes`       | `true`  | Tell axe to run inside iframes                                                                                                          |
-| `elementRef`    | `false` | Return element references in addition to the target                                                                                     |
-| `restoreScroll` | `false` | Scrolls elements back to before axe started                                                                                             |
-| `frameWaitTime` | `60000` | How long (in milliseconds) axe waits for a response from embedded frames before timing out                                              |
-| `preload`       | `false` | Any additional assets (eg: cssom) to preload before running rules. [See here for configuration details](#preload-configuration-details) |
+| Property           | Default | Description                                                                                                                             |
+| ------------------ | :------ | :-------------------------------------------------------------------------------------------------------------------------------------- |
+| `runOnly`          | n/a     | Limit which rules are executed, based on names or tags                                                                                  |
+| `rules`            | n/a     | Allow customizing a rule's properties (including { enable: false })                                                                     |
+| `reporter`         | `v1`    | Which reporter to use (see [Configuration](#api-name-axeconfigure))                                                                     |
+| `resultTypes`      | n/a     | Limit which result types are processed and aggregated                                                                                   |
+| `xpath`            | `false` | Return xpath selectors for elements                                                                                                     |
+| `absolutePaths`    | `false` | Use absolute paths when creating element selectors                                                                                      |
+| `iframes`          | `true`  | Tell axe to run inside iframes                                                                                                          |
+| `elementRef`       | `false` | Return element references in addition to the target                                                                                     |
+| `restoreScroll`    | `false` | Scrolls elements back to before axe started                                                                                             |
+| `frameWaitTime`    | `60000` | How long (in milliseconds) axe waits for a response from embedded frames before timing out                                              |
+| `preload`          | `false` | Any additional assets (eg: cssom) to preload before running rules. [See here for configuration details](#preload-configuration-details) |
+| `performanceTimer` | `false` | Log rule performance metrics to the console                                                                                             |
 
 ###### Options Parameter Examples