--- a/lib/core/base/rule.js
+++ b/lib/core/base/rule.js
@@ -90,16 +90,49 @@ Rule.prototype.matches = function() {
 /**
  * Selects `HTMLElement`s based on configured selector
  * @param  {Context} context The resolved Context object
+ * @param  {Mixed}   options Options specific to this rule
  * @return {Array}           All matching `HTMLElement`s
  */
-Rule.prototype.gather = function(context) {
+Rule.prototype.gather = function(context, options) {
 	'use strict';
+	const markStart = 'mark_gather_start_' + this.id;
+	const markEnd = 'mark_gather_end_' + this.id;
+	const markHiddenStart = 'mark_isHidden_start_' + this.id;
+	const markHiddenEnd = 'mark_isHidden_end_' + this.id;
+
+	if (options.performanceTimer) {
+		axe.utils.performanceTimer.mark(markStart);
+	}
+
 	var elements = axe.utils.select(this.selector, context);
 	if (this.excludeHidden) {
-		return elements.filter(function(element) {
+		if (options.performanceTimer) {
+			axe.utils.performanceTimer.mark(markHiddenStart);
+		}
+
+		elements = elements.filter(function(element) {
 			return !axe.utils.isHidden(element.actualNode);
 		});
+
+		if (options.performanceTimer) {
+			axe.utils.performanceTimer.mark(markHiddenEnd);
+			axe.utils.performanceTimer.measure(
+				'rule_' + this.id + '#gather_axe.utils.isHidden',
+				markHiddenStart,
+				markHiddenEnd
+			);
+		}
 	}
+
+	if (options.performanceTimer) {
+		axe.utils.performanceTimer.mark(markEnd);
+		axe.utils.performanceTimer.measure(
+			'rule_' + this.id + '#gather',
+			markStart,
+			markEnd
+		);
+	}
+
 	return elements;
 };
 
@@ -155,9 +188,7 @@ Rule.prototype.run = function(context, options, resolve, reject) {
 
 	try {
 		// Matches throws an error when it lacks support for document methods
-		nodes = this.gather(context).filter(node =>
-			this.matches(node.actualNode, node, context)
-		);
+		nodes = gatherAndMatchNodes(this, context, options);
 	} catch (error) {
 		// Exit the rule execution if matches fails
 		reject(new SupportError({ cause: error, ruleId: this.id }));
@@ -166,6 +197,7 @@ Rule.prototype.run = function(context, options, resolve, reject) {
 
 	if (options.performanceTimer) {
 		axe.log(
+			this.id,
 			'gather (',
 			nodes.length,
 			'):',
@@ -217,7 +249,7 @@ Rule.prototype.run = function(context, options, resolve, reject) {
 		axe.utils.performanceTimer.mark(markChecksEnd);
 		axe.utils.performanceTimer.mark(markEnd);
 		axe.utils.performanceTimer.measure(
-			'runchecks_' + this.id,
+			'rule_' + this.id + '#runchecks',
 			markChecksStart,
 			markChecksEnd
 		);
@@ -227,6 +259,38 @@ Rule.prototype.run = function(context, options, resolve, reject) {
 
 	q.then(() => resolve(ruleResult)).catch(error => reject(error));
 };
+
+/**
+ * Selects `HTMLElement`s based on configured selector and filters them based on
+ * the rules matches function
+ * @param  {Rule} rule The rule to check for after checks
+ * @param  {Context} context The resolved Context object
+ * @param  {Mixed}   options Options specific to this rule
+ * @return {Array}           All matching `HTMLElement`s
+ */
+function gatherAndMatchNodes(rule, context, options) {
+	const markMatchesStart = 'mark_matches_start_' + this.id;
+	const markMatchesEnd = 'mark_matches_end_' + this.id;
+
+	let nodes = this.gather(context, options);
+
+	if (options.performanceTimer) {
+		axe.utils.performanceTimer.mark(markMatchesStart);
+	}
+
+	nodes = nodes.filter(node => this.matches(node.actualNode, node, context));
+
+	if (options.performanceTimer) {
+		axe.utils.performanceTimer.mark(markMatchesEnd);
+		axe.utils.performanceTimer.measure(
+			'rule_' + rule.id + '#matches',
+			markMatchesStart,
+			markMatchesEnd
+		);
+	}
+
+	return nodes;
+}
 
 /**
  * Iterates the rule's Checks looking for ones that have an after function