--- a/lib/core/base/rule.js
+++ b/lib/core/base/rule.js
@@ -269,16 +269,16 @@ Rule.prototype.run = function(context, options, resolve, reject) {
  * @return {Array}           All matching `HTMLElement`s
  */
 function gatherAndMatchNodes(rule, context, options) {
-	const markMatchesStart = 'mark_matches_start_' + this.id;
-	const markMatchesEnd = 'mark_matches_end_' + this.id;
+	const markMatchesStart = 'mark_matches_start_' + rule.id;
+	const markMatchesEnd = 'mark_matches_end_' + rule.id;
 
-	let nodes = this.gather(context, options);
+	let nodes = rule.gather(context, options);
 
 	if (options.performanceTimer) {
 		axe.utils.performanceTimer.mark(markMatchesStart);
 	}
 
-	nodes = nodes.filter(node => this.matches(node.actualNode, node, context));
+	nodes = nodes.filter(node => rule.matches(node.actualNode, node, context));
 
 	if (options.performanceTimer) {
 		axe.utils.performanceTimer.mark(markMatchesEnd);