--- a/lib/core/base/rule.js
+++ b/lib/core/base/rule.js
@@ -93,8 +93,7 @@ Rule.prototype.matches = function() {
  * @param  {Mixed}   options Options specific to this rule
  * @return {Array}           All matching `HTMLElement`s
  */
-Rule.prototype.gather = function(context, options) {
-	'use strict';
+Rule.prototype.gather = function(context, options = {}) {
 	const markStart = 'mark_gather_start_' + this.id;
 	const markEnd = 'mark_gather_end_' + this.id;
 	const markHiddenStart = 'mark_isHidden_start_' + this.id;