--- a/lib/core/base/rule.js
+++ b/lib/core/base/rule.js
@@ -187,7 +187,7 @@ Rule.prototype.run = function(context, options, resolve, reject) {
 
 	try {
 		// Matches throws an error when it lacks support for document methods
-		nodes = gatherAndMatchNodes(this, context, options);
+		nodes = this.gatherAndMatchNodes(context, options);
 	} catch (error) {
 		// Exit the rule execution if matches fails
 		reject(new SupportError({ cause: error, ruleId: this.id }));
@@ -267,29 +267,29 @@ Rule.prototype.run = function(context, options, resolve, reject) {
  * @param  {Mixed}   options Options specific to this rule
  * @return {Array}           All matching `HTMLElement`s
  */
-function gatherAndMatchNodes(rule, context, options) {
-	const markMatchesStart = 'mark_matches_start_' + rule.id;
-	const markMatchesEnd = 'mark_matches_end_' + rule.id;
+Rule.prototype.gatherAndMatchNodes = function(context, options) {
+	const markMatchesStart = 'mark_matches_start_' + this.id;
+	const markMatchesEnd = 'mark_matches_end_' + this.id;
 
-	let nodes = rule.gather(context, options);
+	let nodes = this.gather(context, options);
 
 	if (options.performanceTimer) {
 		axe.utils.performanceTimer.mark(markMatchesStart);
 	}
 
-	nodes = nodes.filter(node => rule.matches(node.actualNode, node, context));
+	nodes = nodes.filter(node => this.matches(node.actualNode, node, context));
 
 	if (options.performanceTimer) {
 		axe.utils.performanceTimer.mark(markMatchesEnd);
 		axe.utils.performanceTimer.measure(
-			'rule_' + rule.id + '#matches',
+			'rule_' + this.id + '#matches',
 			markMatchesStart,
 			markMatchesEnd
 		);
 	}
 
 	return nodes;
-}
+};
 
 /**
  * Iterates the rule's Checks looking for ones that have an after function