--- a/lib/core/utils/performance-timer.js
+++ b/lib/core/utils/performance-timer.js
@@ -90,7 +90,12 @@ utils.performanceTimer = (function() {
 				window.performance &&
 				window.performance.getEntriesByType !== undefined
 			) {
-				var measures = window.performance.getEntriesByType('measure');
+				// only output measures that were started after axe started, otherwise
+				// we get measures made by the page before axe ran (which is confusing)
+				var axeStart = window.performance.getEntriesByName('mark_axe_start')[0];
+				var measures = window.performance
+					.getEntriesByType('measure')
+					.filter(measure => measure.startTime >= axeStart.startTime);
 				for (var i = 0; i < measures.length; ++i) {
 					var req = measures[i];
 					if (req.name === measureName) {