Add metrics for `rule.gather`, `rule.matches` and `axe.utils.isHidden`. This allows us to see where a rule is taking the most time better. Also rename rules to be sortable based on id. Lastly, have axe only output metrics that were done after axe run, otherwise you get metrics that the website ran before axe did, which makes it really confusing when reading the metric logs.

Example output:

```
Measure rule_aria-valid-attr-value#gather_axe.utils.isHidden took 931ms
Measure rule_aria-valid-attr-value#gather took 931ms
Measure rule_aria-valid-attr-value took 937ms
Measure rule_aria-valid-attr-value#matches took 4ms
Measure rule_aria-valid-attr-value#runchecks took 2ms
Measure rule_aria-valid-attr#gather_axe.utils.isHidden took 960ms
Measure rule_aria-valid-attr#gather took 960ms
Measure rule_aria-valid-attr took 969ms
Measure rule_aria-valid-attr#matches took 9ms
Measure rule_aria-valid-attr#runchecks took 0ms
Measure rule_color-contrasr#gather_axe.utils.isHidden took 0ms
Measure rule_color-contrast#gather took 0ms 
Measure rule_color-contrast#matches took 860ms 
Measure rule_color-contrast took 1264ms 
Measure rule_color-contrast#runchecks took 404ms
```

Closes issue: N/A

## Reviewer checks

**Required fields, to be filled out by PR reviewer(s)**
- [x] Follows the commit message policy, appropriate for next version
- [x] Has documentation updated, a DU ticket, or requires no documentation change
- [x] Includes new tests, or was unnecessary
- [x] Code is reviewed for security by: Stephen